adding a skip link
Hi, 

For people with visual impairment, I've included a skip link that will pop up only if a Tab key is pressed. 
The skip link has is an anchor with an href that redirects to the official RStudio website which I think is more adapted to accessibility issues.  

Best and thank you again for the amazing app. 
Fodil.